feat: divise la page Home en sections
